+++ backend/models/users.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class User (BaseModel):
+    email: str
