+++ frontend/Dockerfile.dev
@@ -1,5 +1,4 @@
 FROM node:18-alpine
-
 # Install Python and essential build tools
 RUN apk add --update --no-cache python3 make g++ && ln -sf python3 /usr/bin/python
 RUN python3 -m ensurepip
@@ -22,11 +21,8 @@ RUN yarn install
 # Copy the rest of our Next.js folder into /app
 COPY . .
 
-# Build the Next.js application
-RUN yarn build
-
 # Ensure port 3000 is accessible to our system
 EXPOSE 3000
 
-# Run yarn start, as we would via the command line 
-CMD ["yarn", "start"]
+# Run yarn dev, as we would via the command line 
+CMD ["yarn", "dev"]
